python bindings

+++ src/lib/cuda_akaze.h
@@ -2,7 +2,7 @@
 #include "cudaImage.h"
 
 float *AllocBuffers(int width, int height, int num, int omax, int &maxpts, std::vector<CudaImage> &buffers, cv::KeyPoint *&pts,
-                    cv::KeyPoint *&ptsbuffer, short *&ptindices, unsigned char *&desc, float *&descbuffer, CudaImage *&ims);
+                    cv::KeyPoint *&ptsbuffer, int *&ptindices, unsigned char *&desc, float *&descbuffer, CudaImage *&ims);
 void InitCompareIndices();
 void FreeBuffers(float *buffers);
 double LowPass(CudaImage &inimg, CudaImage &outimg, CudaImage &temp, double var, int kernsize);
@@ -14,7 +14,7 @@ double Copy(CudaImage &inimg, CudaImage &outimg);
 double ContrastPercentile(CudaImage &img, CudaImage &temp, CudaImage &blur, float perc, int nbins, float &contrast);
 double HessianDeterminant(CudaImage &img, CudaImage &lx, CudaImage &ly, int step);
 double FindExtrema(CudaImage &img, CudaImage &imgp, CudaImage &imgn, float border, float dthreshold, int scale, int octave, float size, cv::KeyPoint *pts, int maxpts);
-void FilterExtrema(cv::KeyPoint *pts, cv::KeyPoint *newpts, short *kptindices, int &nump);
+void FilterExtrema(cv::KeyPoint *pts, cv::KeyPoint *newpts, int *kptindices, int &nump);
 void ClearPoints();
 int GetPoints(std::vector<cv::KeyPoint>& h_pts, cv::KeyPoint *d_pts, int numPts);
 void WaitCuda();
